Add peephole lwz/li reordering
We disable epilogue scheduling in 1.2.5e, but the patched
1.2.5 behavior still has some epilogue scheduling enabled.

When vanilla 1.2.5 and 1.2.5e differ by an lwz/li instruction
swap prior to epilogue register restoration, the vanilla order
seems to actually be the correct one.

Fixes melee's wobj.c and dobj.c matches, so frank can be reenabled for them.