Bugfix: Fix faulty DLL decomp in `PolyObjectMgr::registerClass`.
Before this change, this function would always throw an ERROR during regular operation.  The Ghidra pseudo-code for this function is a mess, so I see how the mistake was made.

Note: Non-matching builds with `PRINT` and `ERROR` are still non-functional rn due to the issue regarding usage of seSystem before `GameCoreSection` initializes it.