SRT: Rename members
We know from a comment in navi.cpp that these members were named this way.  In my view, this fundamental, quasi-POD class is deserving of an exception to our hungarian naming scheme just like Vector3f already has.

It also makes it easier to do what I'm about to do.