Include stdlib headers the correct way
Our codebase is already set up to do this, but we weren't doing it before due to a conflict with math.h and string.h.  Also, we weren't using angle braces which is weird and I don't know fully what might go wrong with that.  Anyway, this is another step toward portability!